+++ sentry-ruby/spec/sentry/hub_spec.rb
@@ -342,6 +342,14 @@ RSpec.describe Sentry::Hub do
       end
     end
 
+    it "reminds users about unsupported options" do
+      expect do
+        subject.capture_event(event, unsupported: true)
+      end.not_to raise_error
+
+      expect(string_io.string).to include("Options [:unsupported] are not supported and will not be applied to the event.")
+    end
+
     context "when event is a transaction" do
       it "transaction.set_context merges and takes precedence over scope.set_context" do
         scope.set_context(:foo, { val: 42 })